Created interfaces of repositories, mocked repositories moved to repositories.imlpMock, added @Builder to all entity classes
Also at mocked repositories instances of objects now created by builder().
