+++ src/main/java/com/epam/aviasales/domain/Account.java
@@ -2,6 +2,7 @@ package com.epam.aviasales.domain;
 
 import java.io.Serializable;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -10,6 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @Getter
 @Setter
+@Builder
 public class Account implements Serializable {
 
   private Long id;
