+++ src/main/java/com/epam/aviasales/domain/Airport.java
@@ -8,6 +8,7 @@ import java.io.Serializable;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class Airport implements Serializable {
 
   private Long id;
