+++ src/main/java/com/epam/aviasales/domain/Flight.java
@@ -1,6 +1,7 @@
 package com.epam.aviasales.domain;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -12,6 +13,7 @@ import java.time.LocalDateTime;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class Flight implements Serializable {
 
   private Long id;
