+++ src/main/java/com/epam/aviasales/domain/PersonalData.java
@@ -2,6 +2,7 @@ package com.epam.aviasales.domain;
 
 import java.io.Serializable;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -12,6 +13,7 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @Getter
 @Setter
+@Builder
 public class PersonalData implements Serializable {
 
   private Long id;
