+++ src/main/java/com/epam/aviasales/services/impl/AccountServiceImpl.java
@@ -2,6 +2,7 @@ package com.epam.aviasales.services.impl;
 
 import com.epam.aviasales.domain.Account;
 import com.epam.aviasales.repositories.AccountRepository;
+import com.epam.aviasales.repositories.implMock.AccountRepositoryImplMock;
 import com.epam.aviasales.services.AccountService;
 import java.util.List;
 import lombok.AccessLevel;
@@ -10,13 +11,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class AccountServiceImpl implements AccountService {
 
-  private static final AccountRepository accountRepository = AccountRepository.getInstance();
   private static final AccountService instance = new AccountServiceImpl();
 
   public static AccountService getInstance() {
     return instance;
   }
 
+  private static final AccountRepository accountRepository = AccountRepositoryImplMock
+      .getInstance();
+
+  @Override
   public List<Account> getAccounts() {
     return accountRepository.getAccounts(1, 20);
   }
