+++ src/main/java/com/epam/aviasales/services/impl/AirplaneServiceImpl.java
@@ -2,6 +2,7 @@ package com.epam.aviasales.services.impl;
 
 import com.epam.aviasales.domain.Airplane;
 import com.epam.aviasales.repositories.AirplaneRepository;
+import com.epam.aviasales.repositories.implMock.AirplaneRepositoryImplMock;
 import com.epam.aviasales.services.AirplaneService;
 import java.util.List;
 import lombok.AccessLevel;
@@ -10,13 +11,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class AirplaneServiceImpl implements AirplaneService {
 
-  private static final AirplaneRepository airplaneRepository = AirplaneRepository.getInstance();
   private static final AirplaneService instance = new AirplaneServiceImpl();
 
   public static AirplaneService getInstance() {
     return instance;
   }
 
+  private static final AirplaneRepository airplaneRepository = AirplaneRepositoryImplMock
+      .getInstance();
+
+  @Override
   public List<Airplane> getAirplanes() {
     return airplaneRepository.getAirplanes(1, 20);
   }
@@ -31,6 +35,7 @@ public class AirplaneServiceImpl implements AirplaneService {
     return airplaneRepository.getByName(name);
   }
 
+  @Override
   public Airplane getById(Long id) {
     return airplaneRepository.getById(id);
   }
