+++ src/main/java/com/epam/aviasales/services/impl/AirportServiceImpl.java
@@ -2,6 +2,7 @@ package com.epam.aviasales.services.impl;
 
 import com.epam.aviasales.domain.Airport;
 import com.epam.aviasales.repositories.AirportRepository;
+import com.epam.aviasales.repositories.implMock.AirportRepositoryImplMock;
 import com.epam.aviasales.services.AirportService;
 import java.util.List;
 import lombok.AccessLevel;
@@ -10,13 +11,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class AirportServiceImpl implements AirportService {
 
-  private static final AirportRepository airportRepository = AirportRepository.getInstance();
   private static final AirportService instance = new AirportServiceImpl();
 
   public static AirportService getInstance() {
     return instance;
   }
 
+  private static final AirportRepository airportRepository = AirportRepositoryImplMock
+      .getInstance();
+
+  @Override
   public List<Airport> getAirports() {
     return airportRepository.getAirports(1, 20);
   }
