+++ src/main/java/com/epam/aviasales/services/impl/PersonalDataServiceImpl.java
@@ -2,6 +2,7 @@ package com.epam.aviasales.services.impl;
 
 import com.epam.aviasales.domain.PersonalData;
 import com.epam.aviasales.repositories.PersonalDataRepository;
+import com.epam.aviasales.repositories.implMock.PersonalDataRepositoryImplMock;
 import com.epam.aviasales.services.PersonalDataService;
 import java.util.List;
 import lombok.AccessLevel;
@@ -10,13 +11,16 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class PersonalDataServiceImpl implements PersonalDataService {
 
-  private static final PersonalDataRepository personalDataRepository = PersonalDataRepository.getInstance();
   private static final PersonalDataService instance = new PersonalDataServiceImpl();
 
   public static PersonalDataService getInstance() {
     return instance;
   }
 
+  private static final PersonalDataRepository personalDataRepository = PersonalDataRepositoryImplMock
+      .getInstance();
+
+  @Override
   public List<PersonalData> getPersonalDatas() {
     return personalDataRepository.getPersonalDatas(1, 20);
   }
